# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Legal Research Platform API with AI capabilities
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Document {
  id: number;
  filename: string;
  originalName?: string;
  fileType: string;
  fileSize: number;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  keywords?: string | null;
  /** @nullable */
  uploadedById?: number | null;
  uploadedAt: string;
  updatedAt?: string;
}

export type DocumentStatsByTypeItem = {
  type: string;
  count: number;
};

export interface DocumentStats {
  total: number;
  byType: DocumentStatsByTypeItem[];
  totalSize: number;
  recentUploads: number;
}

export interface AiSearchInput {
  query: string;
  documentIds?: number[];
  maxResults?: number;
}

export interface SearchHit {
  documentId: number;
  filename: string;
  excerpt: string;
  relevance: number;
}

export interface AiSearchResult {
  query: string;
  results: SearchHit[];
  summary: string;
}

export interface LiteratureReviewInput {
  documentIds: number[];
  topic: string;
  language?: string;
}

export interface LiteratureReviewResult {
  topic: string;
  review: string;
  sources: string[];
  generatedAt?: string;
}

export interface LegalCompareInput {
  topic: string;
  documentIds?: number[];
  aspects?: string[];
}

export interface LegalCompareResult {
  topic: string;
  uaeAnalysis: string;
  franceAnalysis: string;
  similarities: string[];
  differences: string[];
  conclusion: string;
  generatedAt?: string;
}

export interface CitationInput {
  documentId: number;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  publicationYear?: number | null;
  /** @nullable */
  publisher?: string | null;
  /** @nullable */
  url?: string | null;
}

export interface CitationResult {
  citation: string;
  format: string;
}

export interface Comparison {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  rows?: string | null;
  /** @nullable */
  createdById?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface ComparisonInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  rows?: string;
  createdById?: number;
}

export interface ComparisonUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  rows?: string;
}

export interface Comment {
  id: number;
  documentId: number;
  /** @nullable */
  authorId?: number | null;
  /** @nullable */
  authorName?: string | null;
  content: string;
  createdAt: string;
}

export interface CommentInput {
  documentId: number;
  authorId?: number;
  /** @minLength 1 */
  content: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  owner: 'owner',
  supervisor: 'supervisor',
  viewer: 'viewer',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  owner: 'owner',
  supervisor: 'supervisor',
  viewer: 'viewer',
} as const;

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  owner: 'owner',
  supervisor: 'supervisor',
  viewer: 'viewer',
} as const;

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  role?: UserUpdateRole;
}

export interface Settings {
  id: number;
  aiEnabled: boolean;
  maxUploadSizeMb: number;
  allowedFileTypes: string;
  maintenanceMode?: boolean;
  updatedAt: string;
}

export interface SettingsUpdate {
  aiEnabled?: boolean;
  maxUploadSizeMb?: number;
  allowedFileTypes?: string;
  maintenanceMode?: boolean;
}

export type ExportInputType = typeof ExportInputType[keyof typeof ExportInputType];


export const ExportInputType = {
  documents: 'documents',
  comparisons: 'comparisons',
  all: 'all',
} as const;

export interface ExportInput {
  type: ExportInputType;
  documentIds?: number[];
  comparisonIds?: number[];
}

export interface ExportResult {
  downloadUrl: string;
  filename: string;
  generatedAt?: string;
}

export type ListDocumentsParams = {
search?: string;
type?: string;
uploadedBy?: number;
};

export type ListCommentsParams = {
documentId: number;
};