# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Legal Research Platform API with AI capabilities
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/search:
    post:
      operationId: aiSearch
      tags: [ai]
      summary: AI-powered smart search within documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSearchInput"
      responses:
        "200":
          description: AI search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSearchResult"
        "503":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/literature-review:
    post:
      operationId: generateLiteratureReview
      tags: [ai]
      summary: Generate a literature review from selected documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiteratureReviewInput"
      responses:
        "200":
          description: Generated literature review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiteratureReviewResult"
        "503":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/uae-france-compare:
    post:
      operationId: uaeFranceCompare
      tags: [ai]
      summary: AI legal comparison between UAE and France law
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalCompareInput"
      responses:
        "200":
          description: Legal comparison result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalCompareResult"
        "503":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /comparisons:
    get:
      operationId: listComparisons
      tags: [comparisons]
      summary: List manual comparison tables
      responses:
        "200":
          description: List of comparisons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comparison"
    post:
      operationId: createComparison
      tags: [comparisons]
      summary: Create a new comparison table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComparisonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
  /comparisons/{id}:
    get:
      operationId: getComparison
      tags: [comparisons]
      summary: Get a comparison table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comparison table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateComparison
      tags: [comparisons]
      summary: Update a comparison table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComparisonUpdate"
      responses:
        "200":
          description: Updated comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
    delete:
      operationId: deleteComparison
      tags: [comparisons]
      summary: Delete a comparison table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a document
      parameters:
        - name: documentId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AiSearchInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
        documentIds:
          type: array
          items:
            type: integer
        maxResults:
          type: integer
    AiSearchResult:
      type: object
      required: [query, results, summary]
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchHit"
        summary:
          type: string
    SearchHit:
      type: object
      required: [documentId, filename, excerpt, relevance]
      properties:
        documentId:
          type: integer
        filename:
          type: string
        excerpt:
          type: string
        relevance:
          type: number
    LiteratureReviewInput:
      type: object
      required: [documentIds, topic]
      properties:
        documentIds:
          type: array
          items:
            type: integer
        topic:
          type: string
        language:
          type: string
    LiteratureReviewResult:
      type: object
      required: [topic, review, sources]
      properties:
        topic:
          type: string
        review:
          type: string
        sources:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
    LegalCompareInput:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
        documentIds:
          type: array
          items:
            type: integer
        aspects:
          type: array
          items:
            type: string
    LegalCompareResult:
      type: object
      required: [topic, uaeAnalysis, franceAnalysis, similarities, differences, conclusion]
      properties:
        topic:
          type: string
        uaeAnalysis:
          type: string
        franceAnalysis:
          type: string
        similarities:
          type: array
          items:
            type: string
        differences:
          type: array
          items:
            type: string
        conclusion:
          type: string
        generatedAt:
          type: string
          format: date-time
...
    Comparison:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        rows:
          type: ["string", "null"]
        createdById:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ComparisonInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        rows:
          type: string
        createdById:
          type: integer
    ComparisonUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        rows:
          type: string
    Comment:
      type: object
      required: [id, documentId, content, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [documentId, content]
      properties:
        documentId:
          type: integer
        authorId:
          type: integer
        content:
          type: string
          minLength: 1
...