# lib/api-zod/src/generated/types/comparisonInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Legal Research Platform API with AI capabilities
 * OpenAPI spec version: 0.1.0
 */

export interface ComparisonInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  rows?: string;
  createdById?: number;
}