# lib/api-zod/src/generated/types/legalCompareResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Legal Research Platform API with AI capabilities
 * OpenAPI spec version: 0.1.0
 */

export interface LegalCompareResult {
  topic: string;
  uaeAnalysis: string;
  franceAnalysis: string;
  similarities: string[];
  differences: string[];
  conclusion: string;
  generatedAt?: Date;
}